+++ Runtime/!Core/Attributes/Layer/LayerAttibute.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Shadowpaw {
+  /// <summary>
+  /// Attribute to draw an integer field as a Layer.
+  /// This is preferable to LayerMask when you only want to select a single Layer.
+  /// </summary>
+  public class LayerAttribute : Attribute { }
+}