+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Initial release of Shadowpaw - Unity Toolkit
 
 - `License`: Changed to the [UNLICENSE](https://unlicense.org) (Public Domain)
 - `Core`: Added some magic code to enable the init and record keywords in Unity
+- `Attributes`: Added Layer attribute for selecting a singular layer
 - `Utils`: Added extensions for LayerMask, List, Reflection, String, Task, and Type
 
 Be good to each other 💕